refactoring following regis' comments
